NS2.0 Complete FileNotFoundFoundException surface area by sharing it.
This is the CoreRT half of https://github.com/dotnet/coreclr/pull/10775

This also replaces "SetErrorCode()" with "HResult =" to match the
prevailing practice.
